+++ lib/modules/manager/cocoapods/extract.spec.ts
@@ -3,8 +3,6 @@ import { GlobalConfig } from '../../../config/global';
 import type { RepoGlobalConfig } from '../../../config/types';
 import { extractPackageFile } from '.';
 
-jest.mock('../../../util/fs/access');
-
 const simplePodfile = Fixtures.get('Podfile.simple');
 const complexPodfile = Fixtures.get('Podfile.complex');
