+++ lib/modules/manager/flux/extract.spec.ts
@@ -4,8 +4,6 @@ import type { RepoGlobalConfig } from '../../../config/types';
 import type { ExtractConfig } from '../types';
 import { extractAllPackageFiles, extractPackageFile } from '.';
 
-jest.mock('../../../util/fs/access');
-
 const config: ExtractConfig = {};
 const adminConfig: RepoGlobalConfig = { localDir: '' };
