+++ lib/modules/manager/gitlabci/extract.spec.ts
@@ -9,8 +9,6 @@ import {
 } from './extract';
 import { extractAllPackageFiles, extractPackageFile } from '.';
 
-jest.mock('../../../util/fs/access');
-
 const config: ExtractConfig = {};
 
 const adminConfig: RepoGlobalConfig = { localDir: '' };
